+++ src/main/resources/db/migration/V4__create_pedido_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE pedido (
     funcionario_id UUID NOT NULL,
     hora_pedido TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     hora_cancelamento TIMESTAMP,
-    valor_pedido NUMERIC(10, 2) DEFAULT 0.0,
+    valor_pedido DOUBLE PRECISION DEFAULT 0.0,
     FOREIGN KEY (mesa_id) REFERENCES mesa (id) ON DELETE CASCADE,
     FOREIGN KEY (funcionario_id) REFERENCES funcionario (id) ON DELETE SET NULL
 );
